requirements.txt
aniso8601==8.0.0
boto3==1.10.45
botocore==1.13.45
certifi==2019.11.28
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
Flask-RESTful==0.3.7
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.17.3
oauthlib==3.1.0
pandas==0.25.3
py4j==0.10.7
PySocks==1.7.1
python-dateutil==2.8.1
python-twitter==3.5
pytorch-nlp==0.5.0
pytorch-pretrained-bert==0.6.2
pytz==2019.3
regex==2019.12.20
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
sacremoses==0.0.35
sentencepiece==0.1.85
six==1.13.0
torch==1.3.1
tqdm==4.41.0
transformers==2.3.0
tweepy==3.8.0
urllib3==1.25.7
Werkzeug==0.16.0
